Move pipeline, item and spider to their own files
