+++ webcomix/comic_page.py
@@ -1,0 +1,6 @@
+import scrapy
+
+
+class ComicPage(scrapy.Item):
+    image_element = scrapy.Field()
+    page = scrapy.Field()
